package/nut: don't build in parallel
The conditions are not trivial to reproduce, but it can happen that the
headers are not fully regenerated by the time they are included.

This only happens when a python and/or perl interpreter are available,
with a high number of jobs and a high load.

Fixes:
    http://autobuild.buildroot.org/results/523/5231ff39a0839ec5e1962662004214d4b0773068/
    http://autobuild.buildroot.org/results/5a7/5a75d44c028e77b58f0fd9ab794952f2b477dd84/

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Peter Korsgaard <peter@korsgaard.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
(cherry picked from commit 1d8de10c5fb36619708898a529977058886f31d1)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>