nodejs: security bump to version 6.12.2
Fixes CVE-2017-15896 - Node.js was affected by OpenSSL vulnerability
CVE-2017-3737 in regards to the use of SSL_read() due to TLS handshake
failure.  The result was that an active network attacker could send
application data to Node.js using the TLS or HTTP2 modules in a way that
bypassed TLS authentication and encryption.

For more details, see the announcement:
https://nodejs.org/en/blog/vulnerability/december-2017-security-releases/

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>