[wpewebkit] Add BR2_PACKAGE_WPEWEBKIT_BUILD_DEVELOPMENT_VERSION option
Add the BR2_PACKAGE_WPEWEBKIT_BUILD_DEVELOPMENT_VERSION option that
switches to building the in-development version of WPEWebKit (which
is the master branch of that repository) when enabled.

Option is disabled by default. In that case, an older version of the
master branch is used. This non-development version should probably
come from the stable branch.