grub2: separate target and host builds
grub2 requires the host grub2-mkimage tool to build some of its target
images. The current way of building this tool in the grub2 package is to
perform a simultaneous host-tools/target-bootloader build during the
grub2 build step.
This method makes the recipe complex to understand, and proved to be a
complication during work to enable grub2 support for other architectures
than x86.
This patch tries to do a better separation between the build of grub2
host tools and target boot loader image, as a partial step to enable
grub2 to build for other architectures.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>