grub2: enable support for arm and aarch64 targets
This commit enables the arm-uboot, arm-efi and aarch64-efi grub2
platforms in Buildroot.

With the uboot platform, the grub2 image gets built as a u-boot image
and is loaded from u-boot through a regular "bootm". The only
requirement from the u-boot side in order to allow this is that u-boot
is built with CONFIG_API enabled. CONFIG_API seems to not be enabled by
default in most in-tree configurations, however, it seems to be
available for quite some time now. So it might be possible to use this
even on older u-boot versions. This is available only for arm (32-bit).

With the efi platform, grub2 gets built as an EFI executable. This
allows EFI firmware to find and load it similarly as it can be done for
x86_64. Also, since u-boot v2016.05, u-boot is able to load and boot an
EFI executable, so the uboot efi platform can also be used from u-boot
in recent versions. This has been enabled (mostly) by default for ARM
u-boot. efi platform is available for both arm and aarch64.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>