xserver_xorg-server: backport upstream fix for CVE-2018-14665 to 1.19.6
Incorrect command-line parameter validation in the Xorg X server can
lead to privilege elevation and/or arbitrary files overwrite, when the
X server is running with elevated privileges (ie when Xorg is
installed with the setuid bit set and started by a non-root user).

The -modulepath argument can be used to specify an insecure path to
modules that are going to be loaded in the X server, allowing to
execute unprivileged code in the privileged process.

The -logfile argument can be used to overwrite arbitrary files in the
file system, due to incorrect checks in the parsing of the option.

For more details, see the advisory:

https://lists.x.org/archives/xorg-announce/2018-October/002927.html

Issue was introduced in 1.19.0, so the older xserver variants are not
affected.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>