mariadb: drop my-small.cnf handling
Unbreaks builds without BR2_PACKAGE_MARIADB_SERVER as this only gets
installed if the server is enabled.

As pointed out in commit 2b82e014b40 (package/mariadb: bump version to
10.3.10), this file has been removed upstream in newer versions as it hasn't
been updated in >8 years and the compiled in defaults are sensible, so
completely remove the file handling instead of adding logic to only install
it if the server is enabled.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>