package/weston: fix build with latest freerdp
Commit 942da943f3 (package/freerdp: security bump to version 2.0.0-rc4)
backported a patch from master, that bumps FreeRDP for security fixes.

However, the FreeRDP development process is, well, a bit weird, as they
have been in the rc phase for the past two years now (2.0.0-rc0 was on
2017-07-24). As such, they still continue to do huge code refactoring
and cleanup and stuff, even though this is the rc phase. Their API thus
changed quite a bit between rc2 and rc4.

Thus the backport now breaks weston.

We do not have the problem on master or 2019.05.x, because we are using
a more recent weston (6.0.0) which has the fix already. But on 2019.02.x,
we're still using an older weston (5.0.0), which misses the fix.

Backport the fix from upstream.

Fixes (2019.02.x):
    http://autobuild.buildroot.net/results/5fb140acadf26b7f38d3f5c07a4f8dd4dc9acb9e
    http://autobuild.buildroot.org/results/b16610516a8514cf45860a48d693dd0db275e8d9

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
Cc: Alexey Lukyanchuk <skif@skif-web.ru>
Cc: Peter Korsgaard <peter@korsgaard.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>